Bugfix: flash fails if __file__ starts with '../'
The flash script tries to determine the path to itself, but this
failed if the __file__ constant starts with '../'. This happened
because of an extra chdir() at the top of the script that was not
accounted for.

As the __file__ variable is a constant,
os.path.abspath(__file__) assumes the working directory
does not change.