Only use gdb -tui with non-python version
The latest arm-none-eabi-gdb has no -tui support.
When using this latest version, it is recommended to use
arm-none-eabi-gdb-py with a custom .gdbinit file instead,
for example https://github.com/cyrus-and/gdb-dashboard.

When the used gdb ends with -py, gdb is run without the -tui flag.
This should help with backwards compatibility for projects using
the old toolchain.